// src/modules/fpl/dtos/type-fpl-bootstrap.dto.ts
export class TypeFplBootstrapDto {
  events: {
    id: number;
    name: string;
    deadline_time: Date;
    average_entry_score: number;
    finished: boolean;
    data_checked: boolean;
    highest_scoring_entry: number;
    deadline_time_epoch: number;
    deadline_time_game_offset: number;
    highest_score: number;
    is_previous: boolean;
    is_current: boolean;
    is_next: boolean;
    cup_leagues_created: boolean;
    h2h_ko_matches_created: boolean;
    chip_plays: { chip_name: string; num_played: number }[];
    most_selected: number;
    most_transferred_in: number;
    top_element: number;
    top_element_info: { id: number; points: number };
    transfers_made: number;
    most_captained: number;
    most_vice_captained: number;
  }[];
  game_settings: any;
  phases: {
    id: number;
    name: string;
    start_event: number;
    stop_event: number;
  }[];
  teams: {
    code: number;
    draw: number;
    form: null;
    id: number;
    loss: number;
    name: string;
    played: number;
    points: number;
    position: number;
    short_name: string;
    strength: number;
    team_division: null;
    unavailable: boolean;
    win: number;
    strength_overall_home: number;
    strength_overall_away: number;
    strength_attack_home: number;
    strength_attack_away: number;
    strength_defence_home: number;
    strength_defence_away: number;
    pulse_id: number;
  }[];
  total_players: number;
  elements: {
    chance_of_playing_next_round: number;
    chance_of_playing_this_round: number;
    code: number;
    cost_change_event: number;
    cost_change_event_fall: number;
    cost_change_start: number;
    cost_change_start_fall: number;
    dreamteam_count: number;
    element_type: number;
    ep_next: string;
    ep_this: string;
    event_points: number;
    first_name: string;
    form: string;
    id: 1;
    in_dreamteam: false;
    news: string;
    news_added: Date;
    now_cost: number;
    photo: string;
    points_per_game: string;
    second_name: string;
    selected_by_percent: string;
    special: false;
    squad_number: null;
    status: string;
    team: number;
    team_code: number;
    total_points: number;
    transfers_in: number;
    transfers_in_event: number;
    transfers_out: number;
    transfers_out_event: number;
    value_form: string;
    value_season: string;
    web_name: string;
    minutes: number;
    goals_scored: number;
    assists: number;
    clean_sheets: number;
    goals_conceded: number;
    own_goals: number;
    penalties_saved: number;
    penalties_missed: number;
    yellow_cards: number;
    red_cards: number;
    saves: number;
    bonus: number;
    bps: number;
    influence: string;
    creativity: string;
    threat: string;
    ict_index: string;
    influence_rank: number;
    influence_rank_type: number;
    creativity_rank: number;
    creativity_rank_type: number;
    threat_rank: number;
    threat_rank_type: number;
    ict_index_rank: number;
    ict_index_rank_type: number;
    corners_and_indirect_freekicks_order: null;
    corners_and_indirect_freekicks_text: string;
    direct_freekicks_order: null;
    direct_freekicks_text: string;
    penalties_order: null;
    penalties_text: string;
    now_cost_rank: number;
    now_cost_rank_type: number;
    form_rank: number;
    form_rank_type: number;
    points_per_game_rank: number;
    points_per_game_rank_type: number;
    selected_rank: number;
    selected_rank_type: number;
  }[];
  element_stats: { label: string; name: string }[];
  element_types: {
    id: number;
    plural_name: 'Goalkeepers' | 'Defenders' | 'Midfielders' | 'Forwards';
    plural_name_short: string;
    singular_name: string;
    singular_name_short: string;
    squad_select: number;
    squad_min_play: number;
    squad_max_play: number;
    ui_shirt_specific: true;
    sub_positions_locked: number[];
    element_count: number;
  }[];
}